clean conftest
